+++ packages/edition-slices/src/slices/dailyregisterleadfour/index.js
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import { ItemRowSeparator } from "@times-components/slice-layout";
 import { ResponsiveSlice } from "../shared";
 import { TileS } from "../../tiles";
-import styleFactory from "./styles";
+import styleFactory, { backgroundColour } from "./styles";
 import Logo from "./logo";
 
 class DailyRegisterLeadFour extends Component {
@@ -114,12 +114,14 @@ class DailyRegisterLeadFour extends Component {
 
   render() {
     return (
-      <ResponsiveSlice
-        renderHuge={this.renderHuge}
-        renderMedium={this.renderSlice}
-        renderSmall={this.renderSlice}
-        renderWide={this.renderSlice}
-      />
+      <View style={backgroundColour}>
+        <ResponsiveSlice
+          renderHuge={this.renderHuge}
+          renderMedium={this.renderSlice}
+          renderSmall={this.renderSlice}
+          renderWide={this.renderSlice}
+        />
+      </View>
     );
   }
 }