+++ packages/sticky/src/context.js
@@ -6,14 +6,15 @@ import PropTypes from "prop-types";
 import getDisplayName from "react-display-name";
 import hoistNonReactStatic from "hoist-non-react-statics";
 
-import { getTop } from "./util";
+import { getTopFromBody } from "./util";
 
-const { Provider, Consumer } = React.createContext({ top: 0 });
+export const defaultContext = { top: 0 };
+const { Provider, Consumer } = React.createContext(defaultContext);
 
 class StickyProvider extends Component {
   constructor(props) {
     super(props);
-    this.state = { top: 0 };
+    this.state = { ...defaultContext };
     this.ref = this.ref.bind(this);
     this.setTop = this.setTop.bind(this);
   }
@@ -34,7 +35,7 @@ class StickyProvider extends Component {
       return;
     }
 
-    const newTop = Math.round((getTop(node) || 0) + window.pageYOffset);
+    const newTop = Math.round(getTopFromBody(node));
 
     if (newTop !== top) {
       this.setState({ top: newTop });
@@ -59,8 +60,8 @@ class StickyProvider extends Component {
 }
 
 StickyProvider.propTypes = {
-  Component: PropTypes.node,
-  children: PropTypes.element.isRequired
+  Component: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
+  children: PropTypes.node.isRequired
 };
 
 StickyProvider.defaultProps = {