+++ packages/storybook/src/storybook.js
@@ -14,6 +14,8 @@ const {
 
 const showcaseConverter = showcaseToStoryBook(storiesOf, builtInKnobs, actions);
 
+export * from "./styles";
+
 export {
   BarSpacingDecorator,
   CenteredDecorator,