+++ packages/article-lead-asset/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/newsuk/times-components/compare/@times-components/article-lead-asset@1.3.0...@times-components/article-lead-asset@1.3.1) (2019-06-21)
+
+**Note:** Version bump only for package @times-components/article-lead-asset
+
+
+
+
+
 # [1.3.0](https://github.com/newsuk/times-components/compare/@times-components/article-lead-asset@1.2.6...@times-components/article-lead-asset@1.3.0) (2019-06-19)