+++ packages/mocks/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.22](https://github.com/newsuk/times-components/compare/@times-components/mocks@0.0.21...@times-components/mocks@0.0.22) (2019-06-21)
+
+**Note:** Version bump only for package @times-components/mocks
+
+
+
+
+
 ## [0.0.21](https://github.com/newsuk/times-components/compare/@times-components/mocks@0.0.20...@times-components/mocks@0.0.21) (2019-06-18)
 
 **Note:** Version bump only for package @times-components/mocks