+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.77](https://github.com/newsuk/times-components/compare/@times-components/pagination@3.2.76...@times-components/pagination@3.2.77) (2019-06-21)
+
+**Note:** Version bump only for package @times-components/pagination
+
+
+
+
+
 ## [3.2.76](https://github.com/newsuk/times-components/compare/@times-components/pagination@3.2.75...@times-components/pagination@3.2.76) (2019-06-18)
 
 **Note:** Version bump only for package @times-components/pagination