+++ packages/provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.17.12](https://github.com/newsuk/times-components/compare/@times-components/provider@1.17.11...@times-components/provider@1.17.12) (2019-06-21)
+
+**Note:** Version bump only for package @times-components/provider
+
+
+
+
+
 ## [1.17.11](https://github.com/newsuk/times-components/compare/@times-components/provider@1.17.10...@times-components/provider@1.17.11) (2019-06-18)
 
 **Note:** Version bump only for package @times-components/provider