+++ packages/text-flow/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/newsuk/times-components/compare/@times-components/text-flow@0.1.2...@times-components/text-flow@0.1.3) (2019-06-21)
+
+**Note:** Version bump only for package @times-components/text-flow
+
+
+
+
+
 ## [0.1.2](https://github.com/newsuk/times-components/compare/@times-components/text-flow@0.1.1...@times-components/text-flow@0.1.2) (2019-06-18)
 
 **Note:** Version bump only for package @times-components/text-flow