+++ packages/user-state/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.6](https://github.com/newsuk/times-components/compare/@times-components/user-state@0.0.5...@times-components/user-state@0.0.6) (2019-06-21)
+
+**Note:** Version bump only for package @times-components/user-state
+
+
+
+
+
 ## [0.0.5](https://github.com/newsuk/times-components/compare/@times-components/user-state@0.0.4...@times-components/user-state@0.0.5) (2019-06-18)
 
 **Note:** Version bump only for package @times-components/user-state