+++ packages/video/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.46](https://github.com/newsuk/times-components/compare/@times-components/video@4.7.45...@times-components/video@4.7.46) (2019-06-21)
+
+**Note:** Version bump only for package @times-components/video
+
+
+
+
+
 ## [4.7.45](https://github.com/newsuk/times-components/compare/@times-components/video@4.7.44...@times-components/video@4.7.45) (2019-06-19)
 
 **Note:** Version bump only for package @times-components/video